requirements.txt
fastapi[standard]
uvicorn
starlette
python-dotenv